Rename dir 'test' to 'test-env'.
This resolves #144 by fixing a conflict with the `yarn autoclean`
defaults.
