+++ packages/craco/index.js
@@ -12,7 +12,7 @@ const { when, whenDev, whenProd, whenTest } = require("./lib/user-config-utils")
 const { throwUnexpectedConfigError, gitHubIssueUrl } = require("./lib/plugin-utils");
 const { ESLINT_MODES } = require("./lib/features/webpack/eslint");
 const { POSTCSS_MODES } = require("./lib/features/webpack/style/postcss");
-const { createJestConfig } = require("./lib/features/test/api");
+const { createJestConfig } = require("./lib/features/test-env/api");
 const { createWebpackDevConfig, createWebpackProdConfig } = require("./lib/features/webpack/api");
 const { createDevServerConfigProviderProxy } = require("./lib/features/dev-server/api");
